Fix catListTable and friends
The implemtation before would produce incorrect results if tried to get the cartesian product of two queries built out of `unnest`. Which makes sense, because Postgres seems to special case products of `unnest` to have a `ZipList` semantics rather than the normal `[]`.

The solution is to use `evaluate` to rebind the results of such queries, so there is no `unnest` in the expressions we're `<*>`ing.