Try another approach...
We now maintain a stack of bindings from previous subselects in the `Query` monad, which future queries can reference. So for `eval`, to ensure that postgres doesn't try to run a function once where we expect it to be run multiple times, we modify the expression to contain a bunch of superfluous lateral references to the previous queries. This ensures that it gets run every time.